move Occupancy to properties/electronic_eigenvalues.py
